add support for shading
we are only interested in shading svg paths, but we worked on rectangles
during development. the important parts are

- add a Shading dictionary to the page resource dictionary
- add functions to register new entries in the shading dictionary

both of those were just copied from what pdflib does for other
resource types (e.g. XObject).

Next we add shading commands to the drawSVGPath method. if there is a
pattern key, we apply the shading as a fill. it requires ending the path
and setting clip, but other than that seems to work the same way.