fix cmap embedding
although I'm not entirely clear on why this worked, it did work. we add
a default codepoint to every character. apparently when we _use_ the
glpyh we get a valid code point in there. so this just covers the case
where there's no code point.

perhaps there's a more appropriate way to resolve this? ...